Separated netCode readings into two files
